pyproject.toml
[project]
name = "lstm-attention"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "optuna>=4.3.0",
    "pandas>=2.2.3",
    "polars>=1.30.0",
    "rdkit>=2025.3.2",
    "scikit-learn>=1.6.1",
    "pyyaml>=6.0.2",
    "torch==2.6.0+cu118",
    "torchvision==0.21.0+cu118",
    "torch-geometric==2.6.1",
    "torchmetrics>=1.7.1",
    "adjusttext>=1.3.0",
]

[[tool.uv.index]]
name = "torch-cuda"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = { index = "torch-cuda" }
torchvision = { index = "torch-cuda" }

[build-system]
requires = ["setuptools>=67.0.0", "wheel"]
build-backend = "setuptools.build_meta"
